restructure module imports to reduce GUI memory usage
