+++ maestral/gui/folders_dialog.py
@@ -15,7 +15,8 @@ from PyQt5 import QtCore, QtWidgets, QtGui, uic
 from PyQt5.QtCore import QAbstractItemModel, QModelIndex, Qt, QVariant
 
 # maestral modules
-from maestral.sync.main import handle_disconnect, is_child
+from maestral.sync.utils.path import is_child
+from maestral.sync.utils import handle_disconnect
 from maestral.gui.resources import FOLDERS_DIALOG_PATH, get_native_folder_icon
 from maestral.gui.utils import BackgroundTask
