+++ maestral/gui/relink_dialog.py
@@ -15,7 +15,6 @@ from PyQt5 import QtCore, QtWidgets, QtGui, uic
 from PyQt5.QtCore import Qt
 
 # maestral modules
-from maestral.sync.oauth import OAuth2Session
 from maestral.gui.resources import RELINK_DIALOG_PATH, APP_ICON_PATH
 from maestral.gui.utils import get_scaled_font, icon_to_pixmap
 from maestral.gui.utils import BackgroundTask
@@ -25,8 +24,6 @@ logger = logging.getLogger(__name__)
 
 class RelinkDialog(QtWidgets.QDialog):
     """A dialog to show when Maestral's Dropbox access has expired or has been revoked."""
-
-    auth_session = OAuth2Session()
 
     VALID_MSG = "Verified. Restarting Maestral..."
     INVALID_MSG = "Invalid token"
@@ -39,7 +36,11 @@ class RelinkDialog(QtWidgets.QDialog):
         super(self.__class__, self).__init__(parent=None)
         uic.loadUi(RELINK_DIALOG_PATH, self)
 
+        # import OAuth2Session here because of ~40 MB memory footprint
+        from maestral.sync.oauth import OAuth2Session
+
         self._parent = parent
+        self.auth_session = OAuth2Session()
 
         self.setModal(True)
         self.setWindowFlags(Qt.WindowTitleHint | Qt.CustomizeWindowHint)
